refactor predicate function creator
