+++ be/src/exprs/CMakeLists.txt
@@ -62,7 +62,6 @@ add_library(Exprs
   udf_builtins.cpp
   utility_functions.cpp
   info_func.cpp
-  hybrid_set.cpp
   json_functions.cpp
   operators.cpp
   hll_hash_function.cpp
