+++ be/test/exprs/bloom_filter_predicate_test.cpp
@@ -18,6 +18,7 @@
 #include <string>
 
 #include "exprs/bloomfilter_predicate.h"
+#include "exprs/create_predicate_function.h"
 #include "gtest/gtest.h"
 #include "runtime/string_value.h"
 
@@ -32,7 +33,7 @@ public:
 TEST_F(BloomFilterPredicateTest, bloom_filter_func_int_test) {
     auto tracker = MemTracker::CreateTracker();
     std::unique_ptr<IBloomFilterFuncBase> func(
-            IBloomFilterFuncBase::create_bloom_filter(tracker.get(), PrimitiveType::TYPE_INT));
+            create_bloom_filter(tracker.get(), PrimitiveType::TYPE_INT));
     ASSERT_TRUE(func->init(1024, 0.05).ok());
     const int data_size = 1024;
     int data[data_size];
@@ -54,7 +55,7 @@ TEST_F(BloomFilterPredicateTest, bloom_filter_func_int_test) {
 TEST_F(BloomFilterPredicateTest, bloom_filter_func_stringval_test) {
     auto tracker = MemTracker::CreateTracker();
     std::unique_ptr<IBloomFilterFuncBase> func(
-            IBloomFilterFuncBase::create_bloom_filter(tracker.get(), PrimitiveType::TYPE_VARCHAR));
+            create_bloom_filter(tracker.get(), PrimitiveType::TYPE_VARCHAR));
     ASSERT_TRUE(func->init(1024, 0.05).ok());
     ObjectPool obj_pool;
     const int data_size = 1024;
@@ -73,7 +74,7 @@ TEST_F(BloomFilterPredicateTest, bloom_filter_func_stringval_test) {
     ASSERT_FALSE(func->find((const void*)&not_exist_val));
 
     // test fixed char
-    func.reset(IBloomFilterFuncBase::create_bloom_filter(tracker.get(), PrimitiveType::TYPE_CHAR));
+    func.reset(create_bloom_filter(tracker.get(), PrimitiveType::TYPE_CHAR));
     ASSERT_TRUE(func->init(1024, 0.05).ok());
 
     auto varchar_true_str = obj_pool.add(new std::string("true"));
@@ -105,7 +106,7 @@ TEST_F(BloomFilterPredicateTest, bloom_filter_func_stringval_test) {
 TEST_F(BloomFilterPredicateTest, bloom_filter_size_test) {
     auto tracker = MemTracker::CreateTracker();
     std::unique_ptr<IBloomFilterFuncBase> func(
-            IBloomFilterFuncBase::create_bloom_filter(tracker.get(), PrimitiveType::TYPE_VARCHAR));
+            create_bloom_filter(tracker.get(), PrimitiveType::TYPE_VARCHAR));
     int length = 4096;
     func->init_with_fixed_length(4096);
     char* data = nullptr;
