+++ be/test/olap/bloom_filter_column_predicate_test.cpp
@@ -19,6 +19,7 @@
 #include <gtest/gtest.h>
 #include <time.h>
 
+#include "exprs/create_predicate_function.h"
 #include "olap/bloom_filter_predicate.h"
 #include "olap/column_predicate.h"
 #include "olap/field.h"
@@ -91,7 +92,7 @@ TEST_F(TestBloomFilterColumnPredicate, FLOAT_COLUMN) {
 
     auto tracker = MemTracker::CreateTracker(-1, "OlapScanner");
     std::shared_ptr<IBloomFilterFuncBase> bloom_filter(
-            IBloomFilterFuncBase::create_bloom_filter(tracker.get(), PrimitiveType::TYPE_FLOAT));
+            create_bloom_filter(tracker.get(), PrimitiveType::TYPE_FLOAT));
 
     bloom_filter->init(4096, 0.05);
     float value = 4.1;
